Refactor package to reflect application layers.
