+++ src/main/java/pl/kapelinski/mateusz/java/recap/app/bank/api/exception/AccountWithdrawException.java
@@ -1,6 +1,6 @@
-package pl.kapelinski.mateusz.java.recap.app.bank;
+package pl.kapelinski.mateusz.java.recap.app.bank.api.exception;
 
-public class AccountWithdrawException extends Exception {
+public class AccountWithdrawException extends BankException {
 
     public AccountWithdrawException(String message) {
         super(message);
