+++ src/main/java/pl/kapelinski/mateusz/java/recap/app/bank/ui/model/Bank.java
@@ -1,4 +1,6 @@
-package pl.kapelinski.mateusz.java.recap.app.bank;
+package pl.kapelinski.mateusz.java.recap.app.bank.ui.model;
+
+import pl.kapelinski.mateusz.java.recap.app.bank.api.exception.*;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -28,7 +30,7 @@ public class Bank {
         }
     }
 
-    public Client add(Client newClient) throws ClientAddException{
+    public Client add(Client newClient) throws ClientException {
         System.out.println("add(" + newClient + ")");
         if (clients.contains(newClient)) {
            throw new ClientAddException("Already in system");
@@ -57,13 +59,10 @@ public class Bank {
     }
 }
 
-// TODO: 06.05.2022
-//Dodac generowanie uniklanych numerow kont we wlasciwym miejscu (podczas tworzenia klienta)
-//dopsiac reszte w addClient - DONE
 
-// TODO: 11.05.2022
-//Zrobic kilka commitow rozlozonych w czasie z odpowiadajacym opisem. Przeczytac sekcje collaboration - DONE
 
 // TODO: 18.05.2022
-//1. AddClient exception do poprawienie trows golniejszy wyjatek - throw new dokladniejszy wyjatek
-//2. Wszystkie wyjatki rozszerzaja wyjatek BankException
+// dla kazdego modelu dziedziny np. account stworzyc np. AccountRepository, AccountService, AccountController, a modele przeniesc
+
+// TODO: 27.05.2022
+// Dodać refaktoryzacje zmian nazw exception na zwiazane z bankiem
